Added code for serializing pipelines/pipelinesets as json.
Documentation for a sample pipeline in json was added to the doc folder.
I added the whole project responsible for the json code. Currently it has
important code and a dummy-driver just to test it. It requires the forked
version of Visko, and potentially having visko-online and visko in the same directory.