Added php to parse visko objects from json.
viskoapi has some simple classes to wrap around what the system will be getting
from the Java API on the other end. The communication part does not yet do
anything, so a dummy pipelineset in serialized_pipelineset has to be used.

pipeline_adapter has a simple driver to demonstrate how to use the objects.