Added functioning? QueryServlet and supporting classes.
QueryServlet will respond to requests with a QueryResponse, which
contains the pipelineset, and additionally a list of errors. Their
serialization can be gleaned from QueryResponseAdapter.

Also added some initial work on the PipelineJobTable which will support
future execution and status servlets. Joda time library was also added
to the project to support their DateTime objects. This might have messed
up the build path, so check!