Completed (but lightly tested) PipelineExecutionServlet. (/execute)
It turns out you need really need to pass a Pipeline within a
PipelineSet with a valid Query to the API in order to execute.
Theoretically it shouldn't be necessary, but it is (due to weird coding,
and the provenance logging). This means that a complete pipelineset with
Query (and only 1 pipeline) must be passed through the JSON interface.

web/commetest.php also has a short example on how to communicate with
the servlet via PHP. Code needs to be refactored severely before
production use!