Created ViskoBackendCommunicator (ViskoVisualizer).
This class will be responsible for at least the visualization part
of talking to the java backend. Query->pipelineset works but there is
currently a problem with the java servlet that executes pipelines. It keeps
complaining about a missing class (org.mindswap.common.variable)?

commtest.php has example code on how to use ViskoVisualizer