Working DB access for visualization objects.
PipelineManager and QueryManager are responsible for storing
and fetching pipelines and queries respectively. They inherit from abstract class
Manager which is responsible for setting up the database connection.

You must have the correct config.ini in order to connect to the database.

Currently there is very little error handling! So check out querytest.php
to see how to use the managers.