Added support for InputDataURLErrors.
This requires pinging the artifacturl on viskobackend side and
propagating the error upwards.

Also there was previously a bug with Pipelineservlet not setting up the
SPARQL endpoint connection (by not calling parent's init) that prevented
it from working without first using the QueryServlet.

Attempted to add support for MalformedURIErrors but visko has such
absent error handling/distinguishing that I would have to write a total
hack to do so, or otherwise modify the API.